backend/requirements/base.txt
# Django Core
django>=5.0,<5.1
djangorestframework>=3.14.0,<3.15
django-cors-headers>=4.3.0,<4.4
django-guardian>=2.4.0,<2.5
django-filter>=23.5,<24.0

# Authentication
djangorestframework-simplejwt>=5.3.0,<5.4

# Database
psycopg2-binary>=2.9.9,<2.10

# ML/AI Stack - Updated for Python 3.13 compatibility
scikit-learn>=1.5.0
pandas>=2.0.0,<2.3
numpy>=1.26.0,<2.1
joblib>=1.3.0,<1.5

# Task Queue
celery>=5.3.0,<5.4
redis>=5.0.0,<5.1

# API Documentation
drf-spectacular>=0.27.0,<0.28

# Environment Variables
python-dotenv>=1.0.0,<1.1

# Utilities
Pillow>=10.1.0,<10.2
python-dateutil>=2.8.0,<2.9
requests>=2.31.0,<2.32

# Validation
phonenumbers>=8.13.0,<8.14
